Airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Web scraping dependencies
beautifulsoup4
selenium
webdriver-manager
lxml
html5lib

# Data handling
pandas==2.3.2
requests==2.32.5

# Snowflake connector (for pushing scraped data)
snowflake-connector-python==3.17.2
sqlalchemy==1.4.54

# Airflow extras (optional, makes Snowflake hooks/operators available)
apache-airflow-providers-snowflake
